Change logic to dashboardGrain
