+++ OrleansDashboard.Core/DashboardClient.cs
@@ -1,12 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Dynamic;
-using System.Linq;
-using System.Reflection;
+﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using Orleans;
 using Orleans.Concurrency;
-using Orleans.Core;
 using Orleans.Runtime;
 using OrleansDashboard.Model;
 using OrleansDashboard.Model.History;
@@ -76,90 +71,9 @@ namespace OrleansDashboard
             return grainFactory.GetGrain<ISiloGrain>(siloAddress);
         }
 
-        public async Task<ExpandoObject> GetGrainState(string id, string grainType)
+        public async Task<Immutable<string>> GetGrainState(string id, string grainType)
         {
-            var result = new ExpandoObject();
-            try
-            {
-                var implementationType = AppDomain.CurrentDomain.GetAssemblies()
-                                    .SelectMany(s => s.GetTypes())
-                                    .Where(w => w.Name.Equals(grainType))
-                                    .FirstOrDefault();
-
-                var impProperties = implementationType.GetProperties(BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Public);
-
-                var impFields = implementationType.GetFields(BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Public);
-
-
-                var filterProps = impProperties
-                                    .Where(w => w.PropertyType.IsAssignableTo(typeof(IStorage)))
-                                    .Select(s => s.PropertyType.GetGenericArguments().First());
-
-                var filterFields = impFields
-                                    .Where(w => w.FieldType.IsAssignableTo(typeof(IStorage)))
-                                    .Select(s => s.FieldType.GetGenericArguments().First());
-
-
-                var interfaceTypes = implementationType.GetInterfaces();
-
-                foreach (var interfaceType in interfaceTypes)
-                {
-                    try
-                    {
-                        var getGrainMethod = grainFactory.GetType().GetMethods()
-                                             .First( w => w.Name == "GetGrain"
-                                                    && w.ContainsGenericParameters
-                                                    && w.GetParameters().Any(a => a.ParameterType == typeof(GrainId)));
-
-                        var grain = getGrainMethod.MakeGenericMethod(interfaceType).Invoke(grainFactory, new object[] { GrainId.Parse(id.Replace("_", "/")) });
-
-                        var methods = interfaceType.GetMethods()
-                            .Where(w => w.GetParameters().Length == 0
-                            );
-
-                        foreach (var method in methods)
-                        {
-                            try
-                            {
-                                if (method.ReturnType.IsAssignableTo(typeof(Task)) 
-                                    &&
-                                    (
-                                        method.ReturnType.GetGenericArguments()
-                                                    .Any(a => filterProps.Any(f => f == a))
-                                        ||
-                                        method.ReturnType.GetGenericArguments()
-                                                    .Any(a => filterFields.Any(f => f == a))
-                                    )
-                                )
-                                {
-                                    var task = (method.Invoke(grain, null) as Task);
-                                    var resultProperty = task.GetType().GetProperty("Result");
-
-                                    if (resultProperty == null)
-                                        continue;
-
-                                    await task.ConfigureAwait(false);
-
-                                    result.TryAdd(method.Name, resultProperty.GetValue(task));
-                                }
-                            }
-                            catch (Exception ex)
-                            {
-                            }
-                        }
-                    }
-                    catch (Exception ex)
-                    { 
-                    }
-                }
-            }
-            catch (Exception)
-            {
-
-            }
-            
-
-            return result;
+            return await dashboardGrain.GetGrainState(id, grainType);
         }
     }
 }
