+++ OrleansDashboard.Core/IDashboardGrain.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Dynamic;
 using System.Threading.Tasks;
 using Orleans;
 using Orleans.Concurrency;
@@ -24,5 +25,7 @@ namespace OrleansDashboard
         Task<Immutable<Dictionary<string, GrainTraceEntry>>> GetSiloTracing(string address);
 
         Task<Immutable<Dictionary<string, GrainMethodAggregate[]>>> TopGrainMethods(int take);
+
+        Task<Immutable<string>> GetGrainState(string id, string grainType);
     }
 }
