+++ OrleansDashboard.Core/Model/SimpleGrainStatisticCounter.cs
@@ -1,6 +1,4 @@
 ﻿using Orleans;
-using Orleans.Runtime;
-using System;
 
 namespace OrleansDashboard.Model
 {
@@ -33,12 +31,5 @@ namespace OrleansDashboard.Model
 
         [Id(8)]
         public ActivationDetails[] Activations { get; set; }
-    }
-
-    public sealed class ActivationDetails
-    {
-        public GrainId GrainId { get; set; }
-        public Guid? GuidId { get; set; }
-        public long? IntId { get; set; }
     }
 }
